+++ src/OpenTelemetry.Instrumentation.AspNetCore/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Removed `netstandard2.0` and `netstandard2.1` targets
+  ([#3576](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3567))
+
 ## 1.0.0-rc9.5
 
 Released 2022-Aug-02