+++ src/OpenTelemetry.Instrumentation.AspNetCore/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 1.0.0-rc9.6
+
+Released 2022-Aug-18
+
 * Removed `netstandard2.0` and `netstandard2.1` targets. .NET 5 reached EOL
   in May 2022 and .NET Core 3.1 reaches EOL in December 2022. The
   instrumentation for ASP.NET Core now requires .NET 6 or later.